make only one request to create actions
